feat(peerresponsemanager): setup peer manager
Refactor peer manager to handle abstract processes, then use the base class for the peer message
manager and the peer response manager
